requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
python-dotenv==0.17.0
httpx==0.17.1
Faker==8.1.1
model-bakery==1.3.0
